Converted from GLUT windowing to SDL Object oriented structure achieved Rejoice
