+++ CMakeLists.txt
@@ -3,19 +3,25 @@ project(opulence)
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 
+# Use our modified FindSDL2* modules
+set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${opulence_SOURCE_DIR}/CMakeFiles")
+
+# Set an output directory for our binaries
+set(BIN_DIR ${opulence_SOURCE_DIR}/bin)
+
+find_package(SDL2 REQUIRED)
 find_package(OpenGL REQUIRED)
 find_package(GLEW REQUIRED)
-find_package(GLUT REQUIRED)
 
 include_directories(
   ${PROJECT_BINARY_DIR}
   ${OPENGL_INCLUDE_DIR}
+  ${SDL2_INCLUDE_DIR}
   ${GLEW_INCLUDE_DIRS}
-  ${GLUT_INCLUDE_DIRS}
   )
 
-set(SOURCE_FILES opulence.c)
-set_source_files_properties(opulence.c PROPERTIES LANGUAGE CXX )
+set(SOURCE_FILES main.c opulence.c)
+set_source_files_properties(main.c opulence.c PROPERTIES LANGUAGE CXX )
 add_executable(opulence ${SOURCE_FILES})
 
-target_link_libraries(opulence ${OPENGL_LIBRARIES} ${GLEW_LIBRARIES} ${GLUT_LIBRARIES})
+target_link_libraries(opulence ${SDL2_LIBRARY} ${OPENGL_LIBRARIES} ${GLEW_LIBRARIES})
