+++ main.c
@@ -1,0 +1,7 @@
+#include "opulence.c"
+
+int main (int argc, char* argv[]) {
+    Opulence opl;
+    opl.go();
+    return 0;
+}
